requirements.txt
faiss-cpu==1.12.0
google-adk==1.12.0
google-genai==1.31.0
kaleido==1.0.0
lmfit==1.3.4
mp-api==0.45.8
numpy==2.3.2
openai==1.101.0
pandas==2.3.2
plotly==6.3.0
pymatgen==2025.6.14
PyMuPDF==1.26.4
python-dotenv==1.1.1
pydantic==2.11.7
requests==2.32.5
scipy==1.16.1
starlette==0.47.3
uvicorn[standard]==0.34.0
